refactor: type & constant 파일 분리
