+++ packages/my-components/src/DatePicker/DatePickerCalendar/DatePickerCalendar.types.ts
@@ -1,0 +1,3 @@
+import { DateValue, RangeDateValue } from '../DatePicker.types';
+
+export type ModeDate<T> = T extends 'single' ? DateValue : RangeDateValue;
