+++ packages/my-components/src/DatePicker/DatePickerInputs/DatePickerInputs.constants.ts
@@ -1,0 +1,1 @@
+export const DATE_REGEX = new RegExp(/^[0-9./]*$/);
